+++ internal/paths/open_nofollow_windows.go
@@ -7,3 +7,7 @@ import "os"
 func openFileNoFollow(path string, flag int, perm os.FileMode) (*os.File, error) {
 	return os.OpenFile(path, flag, perm)
 }
+
+func syncParentDir(string) error {
+	return nil
+}